Handling error & update algo

+++ app/controller/therapistsController.js
@@ -12,27 +12,51 @@ const therapistsController = {
      * @returns {json} liste des therapists
      */
     async getAll(_,res,next) {
+
         try {
             const allTherapists = await therapistsDatamapper.findAll();
-            res.json(allTherapists)
-        } catch { 
+            if (!allTherapists) {
+                next(new APIError("Mauvaise requête",400))
+
+            } else if (allTherapists.length === 0)  {
+                next(new APIError("La route n'a pas été trouvé",404))
+
+            } else {
+                res.json(allTherapists) 
+            }   
+            } catch { 
             next(new APIError("Erreur lors de la récupération des therapists",500))
         }
     },
     /**
      * Récupération d'un therapist par son id
-     * @param {*}req requête Express
+     * @param {therapists_id}req requête Express
      * @param {*}res réponse Express
      * @returns {json} un therapist
      */
-    async getById(req,res) {
+    async getById(req,res,next) {
         const id = req.params.id
+
+        if (!id) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
         const getTherapistById = await therapistsDatamapper.findByPk(id);
-        res.json(getTherapistById)
-    } catch {
+        
+        if (!getTherapistById) {
+            next(new APIError("Mauvaise requête",400))
+            
+        } else if (getTherapistById.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404));
+        } else {
+            res.json(getTherapistById)  
+        }
+
+        } catch {
         next(new APIError("Erreur lors de la récupération d'un therapists",500))
-    }
+        }
     },
     /**
      * Création d'un therapist
@@ -40,8 +64,7 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} le therapist créé
      */
-    async creatTherapist (req,res) {
-        try {
+    async creatTherapist (req,res,next) {
         const therapistInfo = {
             firstname: req.body.firstname,
             lastname: req.body.lastname,
@@ -54,21 +77,32 @@ const therapistsController = {
             city:req.body.city,
             gender:req.body.gender,
         }
-        const creatTherapist = await therapistsDatamapper.create(therapistInfo);
-        res.json(creatTherapist)
-    } catch {
-        next(new APIError("Erreur lors de la création d'un therapists",500))
+        if (!therapistInfo) {
+            next(new APIError("Paramètres manquants",400));
+            return;
         }
+        try {
+        const creatTherapist = await therapistsDatamapper.create(therapistInfo);
+        if (!creatTherapist) {
+            next(new APIError("Mauvaise requête",400));
+        } else if (creatTherapist.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404))
+        } else {
+            res.json(creatTherapist)
+        }
+        } catch {
+            next(new APIError("Erreur lors de la création d'un therapists",500))
+            }
     },
     /**
      * Mise à jour d'un therapist
-     * @param {*}req requête Express
+     * @param {therapists_id}req requête Express
      * @param {*}res réponse Express
      * @returns {json} le therapist mis à jour
      */
-    async updateTherapist (req,res) {
+    async updateTherapist (req,res,next) {
         const id = req.params.id
-        try {
+
         const therapistInfo = {
             firstname: req.body.firstname,
             lastname: req.body.lastname,
@@ -82,27 +116,58 @@ const therapistsController = {
             gender:req.body.gender,
             role: req.body.role
         }
-        
-        const updateTherapist = await therapistsDatamapper.update({id},therapistInfo)
-        res.json(updateTherapist)
-        } catch {
-            next(new APIError("Erreur lors de la mise à jour d'un therapists",500))
+
+        if (!id) {
+            next(new APIError("Paramètres manquants",400));
+            return;
         }
+        
+        try {
+        const updateTherapist = await therapistsDatamapper.update({id},therapistInfo)
+        
+        if (!updateTherapist) {
+            next(new APIError("Mauvaise requête",400))
+            
+        } else if (updateTherapist.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404));
+
+        } else {
+            res.json(updateTherapist)
+        }
+        } catch {
+        next(new APIError("Erreur lors de la récupération d'un therapists",500))
+        }
+        
     },
     /**
      * Suppression d'un therapists
-     * @param {*}req requête Express
+     * @param {therapist_id}req requête Express
      * @param {*}res réponse Express
      * @returns {json} le therapists supprimé
      */
-    async deleteTherapist (req,res) { 
+    async deleteTherapist (req,res,next) { 
         const id = req.params.id
+
+        if (!id) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
         const deleteTherapist = await therapistsDatamapper.delete(id)
-        res.json(deleteTherapist)
+        if (!deleteTherapist) {
+            next(new APIError("Mauvaise requête",400)) 
+            
+        } else if (deleteTherapist.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404))  
+
+        } else {
+            res.json(deleteTherapist)
+        }
         } catch {
             next(new APIError("Erreur lors de la suppression d'un therapists",500))
         }
+    
     },
     /**
      * Récupération de tous les therapists avec leurs spécialités
@@ -110,11 +175,25 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} liste des therapists avec leurs spécialités
      */
-    async findTherapistsWithSpecialties (req,res) {
+    async findTherapistsWithSpecialties (req,res,next) {
         const id = req.params.id
+
+        if (!id) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
         const findTherapistsWithSpecialties = await therapistsDatamapper.findTherapistsWithSpecialties(id);
+        
+        if(!findTherapistsWithSpecialties) {
+            next(new APIError("Mauvaise requête",400));
+            
+        } else if (findTherapistsWithSpecialties.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404))
+        } else {
         res.json(findTherapistsWithSpecialties)
+        }
         } catch {
             next(new APIError("Erreur lors de la récupération des therapists avec leurs spécialités",500))
         }
@@ -125,12 +204,21 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} liste des therapists avec leurs spécialités
      */
-    async findAllTherapistsWithSpecialities (_,res) {
+    async findAllTherapistsWithSpecialities (_,res,next) {
         try {
         const findAllTherapistsWithSpecialities = await therapistsDatamapper.AllTherapistsWithSpecialities();
-        res.json(findAllTherapistsWithSpecialities)
+        
+        if (!findAllTherapistsWithSpecialities) {
+            next(new APIError("Mauvaise requête",400));
+
+        } else if (findAllTherapistsWithSpecialities.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404))
+            
+        } else {
+            res.json(findAllTherapistsWithSpecialities)
+        }
         } catch {
-            next(new APIError("Erreur lors de la récupération des therapists avec leurs spécialités",500))
+            next(new APIError("Erreur lors de la récupération de tout les therapists avec leurs spécialités",500))
         }
     },
     /**
@@ -139,12 +227,26 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} Ajouter une spécialité à un thérapist
      */
-    async addSpecialtiesToTherapist (req,res) {
+    async addSpecialtiesToTherapist (req,res,next) {
         const therapistId = req.params.therapistId
         const specialityId = req.params.specialityId
+        
+        if(!therapistId || !specialityId) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
         const addSpecialtiesToTherapist = await therapistsDatamapper.addSpecialtiesToTherapist(therapistId, specialityId);
-        res.json(addSpecialtiesToTherapist)
+        if (!addSpecialtiesToTherapist) {
+            next(new APIError("Mauvaise requête",400));
+                        
+        } else if (addSpecialtiesToTherapist.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404))
+            
+        } else {
+            res.json(addSpecialtiesToTherapist)
+        }
         } catch {
             next(new APIError("Erreur lors de l'ajout d'une spécialité à un thérapeute",500))
         }
@@ -155,12 +257,27 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} Supprimer une spécialité à un thérapist
      */
-    async removeSpecialtiesFromTherapist (req,res) {
+    async removeSpecialtiesFromTherapist (req,res,next) {
         const therapistId = req.params.therapistId
         const specialityId = req.params.specialityId
+
+        if(!therapistId || !specialityId) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
         const removeSpecialtiesToTherapist = await therapistsDatamapper.removeSpecialtiesFromTherapist(therapistId, specialityId);
-        res.json(removeSpecialtiesToTherapist)
+        
+        if (!removeSpecialtiesToTherapist) {
+            next(new APIError("Mauvaise requête",400));
+            
+        } else if (removeSpecialtiesToTherapist.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404))
+
+        } else {
+            res.json(removeSpecialtiesToTherapist)
+        }
         } catch {
         next(new APIError("Erreur lors de la suppression d'une spécialité à un thérapeute",500))
         }
@@ -171,11 +288,26 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} liste des therapists avec leurs genres et leurs spécialités
      */
-    async getAllTherapistsByGenderWithSpecialities (req,res) {
+    async getAllTherapistsByGenderWithSpecialities (req,res,next) {
         const gender = req.params.gender
+
+        if (!gender) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
-        const getAllTherapistsByGender = await therapistsDatamapper.getAllTherapistsByGenderWithSpecialities(gender);
-        res.json(getAllTherapistsByGender)
+        const getAllTherapistsByGenderWithSpe = await therapistsDatamapper.getAllTherapistsByGenderWithSpecialities(gender);
+        
+        if (!getAllTherapistsByGenderWithSpe) {
+            next(new APIError("Mauvaise requête",400))
+            
+        } else if (getAllTherapistsByGenderWithSpe.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404));
+            
+        }else {
+            res.json(getAllTherapistsByGenderWithSpe)
+        }
         } catch {
             next(new APIError("Erreur lors de la récupération des therapists par leurs genres et avec leurs spécialités",500))
         }
@@ -186,11 +318,28 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} liste des therapists avec leurs genres
      */
-    async getAllTherapistsByGender (req,res) {
+    async getAllTherapistsByGender (req,res,next) {
         const gender = req.params.gender
+
+        if (!gender) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
         const getAllTherapistsByGender = await therapistsDatamapper.getAllTherapistsByGender(gender);
-        res.json(getAllTherapistsByGender)
+        
+        if (!getAllTherapistsByGender) {
+            next(new APIError("Mauvaise requête",400))
+            
+            
+        } else if (getAllTherapistsByGender.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404));
+            
+        } else {
+            res.json(getAllTherapistsByGender)
+
+        }
         } catch {
             next(new APIError("Erreur lors de la récupération des therapists par leurs genres",500))
         }
@@ -201,11 +350,27 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} liste des rendez-vous d'un therapist
      */
-    async getAllAppointmentOfATherapist (req,res) {
+    async getAllAppointmentOfATherapist (req,res,next) {
         const id = req.params.id;
+
+        if (!id) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
         const getAllAppointmentOfATherapist = await therapistsDatamapper.getAllAppointmentOfATherapist(id);
-        res.json(getAllAppointmentOfATherapist)
+        
+        if (!getAllAppointmentOfATherapist) {
+            next(new APIError("Mauvaise requête",400))
+            
+        } else if (getAllAppointmentOfATherapist.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404));
+
+        } else {
+            res.json(getAllAppointmentOfATherapist)
+            
+        }
         } catch {
             next(new APIError("Erreur lors de la récupération des rendez-vous d'un therapist",500))
         }
@@ -216,13 +381,29 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} un rendez-vous d'un therapist
      */
-    async getOneAppointmentOfATherapist (req,res) {
+    async getOneAppointmentOfATherapist (req,res,next) {
         const therapistId = req.params.therapistId
         const appointmentId = req.params.appointmentId
+
+        if (!therapistId || !appointmentId) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
         const getOneAppointmentOfATherapist = await therapistsDatamapper.getOneAppointmentOfATherapist(therapistId,appointmentId);
-        res.json(getOneAppointmentOfATherapist)
-        } catch {
+        
+        if (!getOneAppointmentOfATherapist) {
+            next(new APIError("Mauvaise requête",400))
+            
+        } else if (getOneAppointmentOfATherapist.length) {
+            next(new APIError("La route n'a pas été trouvé",404));
+
+        }else{
+            res.json(getOneAppointmentOfATherapist)
+
+        }
+         } catch {
             next(new APIError("Erreur lors de la récupération d'un rendez-vous d'un therapist",500))
         }
     },
@@ -232,10 +413,15 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} création de rendez-vous d'un therapist avec un patient
      */
-    async creatAppointmentWithOnePatient (req,res) {
+    async creatAppointmentWithOnePatient (req,res,next) {
         const therapistId = req.params.therapistId
         const patientId= req.params.patientId
-        try {
+
+        if (!therapistId || !patientId) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+        
         const appointment = {
             beginninghour: req.body.beginninghour,
             endtime: req.body.endtime,
@@ -247,8 +433,24 @@ const therapistsController = {
             sessionatoffice : req.body.sessionatoffice,
         }
         
+        if (!appointment) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
+        try {
         const createAppointmentOneTherapist = await therapistsDatamapper.creatAppointmentWithOnePatient({therapistId,patientId},appointment);
-        res.json(createAppointmentOneTherapist);
+        
+        if (!createAppointmentOneTherapist) {
+            next(new APIError("Mauvaise requête",400));
+            
+        } else if (!createAppointmentOneTherapist.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404))
+            
+        } else {
+            res.json(createAppointmentOneTherapist);
+
+        }
         } catch {
         next(new APIError("Erreur lors de la création d'un rendez-vous d'un therapist avec un patient",500))
         }
@@ -259,11 +461,27 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} un avis sur un therapist
      */
-    async viewOneTherapistReviews(req,res) {
+    async viewOneTherapistReviews(req,res,next) {
         const id = req.params.id
+
+        if (!id) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
+
         try {
         const viewOneTherapistReviews = await therapistsDatamapper.viewOneTherapistReviews(id)
-        res.json(viewOneTherapistReviews)
+        
+        if(!viewOneTherapistReviews) {
+            next(new APIError("Mauvaise requête",400));
+
+        } else if (viewOneTherapistReviews.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404))
+
+        } else {
+            res.json(viewOneTherapistReviews)
+            
+        }
         } catch {
             next(new APIError("Erreur lors de la récupération d'un avis sur un therapist",500))
         }
@@ -274,11 +492,26 @@ const therapistsController = {
      * @param {*}res réponse Express
      * @returns {json} tous les therapists par leurs spécialités
      */
-    async findAllTherapistBySpecialties (req,res) {
+    async findAllTherapistBySpecialties (req,res,next) {
         const id = req.params.id
+
+        if (!id) {
+            next(new APIError("Paramètres manquants",400));
+            return;
+        }
         try {
         findAllTherapistBySpecialties = await therapistsDatamapper.findAllTherapistBySpecialties(id)
-        res.json(findAllTherapistBySpecialties)
+        
+        if (!findAllTherapistBySpecialties) {
+            next(new APIError("Mauvaise requête",400));
+                       
+        } else if(findAllTherapistBySpecialties.length === 0) {
+            next(new APIError("La route n'a pas été trouvé",404))
+            
+        }else {
+            res.json(findAllTherapistBySpecialties)
+            
+        }
         } catch {
             next(new APIError("Erreur lors de la récupération de tous les therapists par leurs spécialités",500))
         }
